+++ python/openinference-instrumentation/src/openinference/instrumentation/config.py
@@ -1,29 +1,20 @@
-import collections.abc
-import inspect
 import json
 import os
 from collections.abc import Mapping, Sequence
-from contextlib import contextmanager
 from dataclasses import asdict, dataclass, field, fields
 from datetime import datetime
 from json import JSONEncoder
-from secrets import randbits
 from types import ModuleType, TracebackType
-from typing import (  #  type: ignore[attr-defined]
+from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
     Dict,
-    Iterator,
-    Literal,
     Optional,
     Tuple,
     Type,
-    TypeVar,
     Union,
-    _TypedDictMeta,
     get_args,
-    get_origin,
 )
 
 from opentelemetry.context import (
@@ -32,15 +23,8 @@ from opentelemetry.context import (
     detach,
     set_value,
 )
-from opentelemetry.sdk.trace.id_generator import IdGenerator
-from opentelemetry.trace import (
-    INVALID_SPAN_ID,
-    INVALID_TRACE_ID,
-    Status,
-    StatusCode,
-)
 from opentelemetry.util.types import AttributeValue
-from typing_extensions import ParamSpec, TypeAlias, TypeGuard, _AnnotatedAlias
+from typing_extensions import TypeGuard
 
 from openinference.semconv.trace import (
     EmbeddingAttributes,
@@ -52,42 +36,18 @@ from openinference.semconv.trace import (
     SpanAttributes,
 )
 
+from ._types import OpenInferenceMimeType, OpenInferenceSpanKind
 from .logging import logger
 
 if TYPE_CHECKING:
     from _typeshed import DataclassInstance
 
-    from ._spans import OpenInferenceSpan
-    from ._tracers import OITracer
 
 pydantic: Optional[ModuleType]
 try:
     import pydantic  # try to import without adding a dependency
 except ImportError:
     pydantic = None
-
-
-OpenInferenceMimeType: TypeAlias = Union[
-    Literal["application/json", "text/plain"],
-    OpenInferenceMimeTypeValues,
-]
-OpenInferenceSpanKind: TypeAlias = Union[
-    Literal[
-        "agent",
-        "chain",
-        "embedding",
-        "evaluator",
-        "guardrail",
-        "llm",
-        "reranker",
-        "retriever",
-        "tool",
-        "unknown",
-    ],
-    OpenInferenceSpanKindValues,
-]
-ParametersType = ParamSpec("ParametersType")
-ReturnType = TypeVar("ReturnType")
 
 
 class suppress_tracing:
@@ -384,12 +344,7 @@ class TraceConfig:
             return cast_to(value)
 
 
-_IMPORTANT_ATTRIBUTES = [
-    SpanAttributes.OPENINFERENCE_SPAN_KIND,
-]
-
-
-def get_span_kind_attributes(kind: OpenInferenceSpanKind, /) -> Dict[str, AttributeValue]:
+def get_span_kind_attributes(kind: "OpenInferenceSpanKind", /) -> Dict[str, AttributeValue]:
     normalized_kind = _normalize_openinference_span_kind(kind)
     return {
         OPENINFERENCE_SPAN_KIND: normalized_kind.value,
@@ -510,23 +465,6 @@ def get_tool_attributes(
     return attributes
 
 
-class _IdGenerator(IdGenerator):
-    """
-    An IdGenerator that uses a different source of randomness to
-    avoid being affected by seeds set by user application.
-    """
-
-    def generate_span_id(self) -> int:
-        while (span_id := randbits(64)) == INVALID_SPAN_ID:
-            continue
-        return span_id
-
-    def generate_trace_id(self) -> int:
-        while (trace_id := randbits(128)) == INVALID_TRACE_ID:
-            continue
-        return trace_id
-
-
 def is_base64_url(url: str) -> bool:
     if not isinstance(url, str):
         return False
@@ -542,7 +480,9 @@ def _normalize_mime_type(mime_type: OpenInferenceMimeType) -> OpenInferenceMimeT
         raise ValueError(f"Invalid mime type: {mime_type}")
 
 
-def _normalize_openinference_span_kind(kind: OpenInferenceSpanKind) -> OpenInferenceSpanKindValues:
+def _normalize_openinference_span_kind(
+    kind: "OpenInferenceSpanKind",
+) -> OpenInferenceSpanKindValues:
     if isinstance(kind, OpenInferenceSpanKindValues):
         return kind
     if not kind.islower():
@@ -563,262 +503,6 @@ def _is_dataclass_instance(obj: Any) -> TypeGuard["DataclassInstance"]:
     cls = type(obj)
     return hasattr(cls, "__dataclass_fields__")
 
-
-class _ChainContext:
-    def __init__(self, span: "OpenInferenceSpan") -> None:
-        self._span = span
-
-    def process_output(self, output: ReturnType) -> ReturnType:
-        attributes = getattr(self._span, "attributes", {})
-        has_output = OUTPUT_VALUE in attributes
-        if not has_output:
-            self._span.set_output(value=output)
-        return output
-
-
-@contextmanager
-def _chain_context(
-    *,
-    tracer: "OITracer",
-    name: Optional[str],
-    kind: OpenInferenceSpanKindValues,
-    wrapped: Callable[ParametersType, ReturnType],
-    instance: Any,
-    args: Tuple[Any, ...],
-    kwargs: Dict[str, Any],
-) -> Iterator[_ChainContext]:
-    span_name = name or _infer_span_name(instance=instance, callable=wrapped)
-    bound_args = inspect.signature(wrapped).bind(*args, **kwargs)
-    bound_args.apply_defaults()
-    arguments = bound_args.arguments
-
-    if len(arguments) == 1:
-        argument = next(iter(arguments.values()))
-        input_attributes = get_input_attributes(argument)
-    else:
-        input_attributes = get_input_attributes(arguments)
-
-    with tracer.start_as_current_span(
-        span_name,
-        openinference_span_kind=kind,
-        attributes=input_attributes,
-    ) as span:
-        context = _ChainContext(span=span)
-        yield context
-        span.set_status(Status(StatusCode.OK))
-
-
-class _ToolContext:
-    def __init__(self, span: "OpenInferenceSpan") -> None:
-        self._span = span
-
-    def process_output(self, output: ReturnType) -> ReturnType:
-        attributes = getattr(self._span, "attributes", {})
-        has_output = OUTPUT_VALUE in attributes
-        if not has_output:
-            self._span.set_output(value=output)
-        return output
-
-
-@contextmanager
-def _tool_context(
-    *,
-    tracer: "OITracer",
-    name: Optional[str],
-    description: Optional[str],
-    parameters: Optional[Union[str, Dict[str, Any]]],
-    wrapped: Callable[ParametersType, ReturnType],
-    instance: Any,
-    args: Tuple[Any, ...],
-    kwargs: Dict[str, Any],
-) -> Iterator[_ToolContext]:
-    tool_name = name or _infer_span_name(instance=instance, callable=wrapped)
-    bound_args = inspect.signature(wrapped).bind(*args, **kwargs)
-    bound_args.apply_defaults()
-    arguments = bound_args.arguments
-    input_attributes = get_input_attributes(arguments)
-    tool_description = description or _infer_tool_description(wrapped)
-    tool_parameters = parameters or _infer_tool_parameters(
-        callable=wrapped,
-        tool_name=tool_name,
-        tool_description=tool_description,
-    )
-    tool_attributes = get_tool_attributes(
-        name=tool_name,
-        description=tool_description,
-        parameters=tool_parameters,
-    )
-    with tracer.start_as_current_span(
-        tool_name,
-        openinference_span_kind=OpenInferenceSpanKindValues.TOOL,
-        attributes={
-            **input_attributes,
-            **tool_attributes,
-        },
-    ) as span:
-        context = _ToolContext(span=span)
-        yield context
-        span.set_status(Status(StatusCode.OK))
-
-
-def _infer_span_name(*, instance: Any, callable: Callable[..., Any]) -> str:
-    """
-    Makes a best-effort attempt to infer a span name from the bound instance
-    (e.g., self or cls) and the callable (the function or method being wrapped).
-    Handles functions, methods, and class methods.
-    """
-
-    if inspect.ismethod(callable):
-        is_class_method = isinstance(instance, type)
-        if is_class_method:
-            class_name = instance.__name__
-        else:  # regular method
-            class_name = instance.__class__.__name__
-        method_name = callable.__name__
-        return f"{class_name}.{method_name}"
-    function_name = callable.__name__
-    return function_name
-
-
-def _infer_tool_description(callable: Callable[..., Any]) -> Optional[str]:
-    """
-    Infers a tool description from the callable's docstring if one exists.
-    """
-    docstring = callable.__doc__
-    if docstring is not None and (stripped_docstring := docstring.strip()):
-        return stripped_docstring
-    return None
-
-
-def _infer_tool_parameters(
-    *,
-    callable: Callable[..., Any],
-    tool_name: str,
-    tool_description: Optional[str],
-) -> Dict[str, Any]:
-    json_schema: Dict[str, Any] = {"type": "object", "title": tool_name}
-    if tool_description:
-        json_schema["description"] = tool_description
-    properties = {}
-    required_properties = []
-    signature = inspect.signature(callable)
-    for param_name, param in signature.parameters.items():
-        property_data = {}
-        default_value = param.default
-        has_default = default_value is not inspect.Parameter.empty
-        if has_default:
-            property_data["default"] = default_value
-        if not has_default:
-            required_properties.append(param_name)
-        annotation = param.annotation
-        property_data.update(_get_jsonschema_type(annotation))
-        metadata = getattr(annotation, "__metadata__", None)
-        description: Optional[str] = None
-        if metadata and isinstance(first_metadata := metadata[0], str):
-            description = first_metadata
-        if description:
-            property_data["description"] = description
-        properties[param_name] = property_data
-
-    json_schema["properties"] = properties
-    if required_properties:
-        json_schema["required"] = required_properties
-    return json_schema
-
-
-def _get_jsonschema_type(annotation_type: type) -> Dict[str, Any]:
-    if isinstance(annotation_type, _AnnotatedAlias):
-        annotation_type = annotation_type.__args__[0]
-        return _get_jsonschema_type(annotation_type)
-    if annotation_type is type(None) or annotation_type is None:
-        return {"type": "null"}
-    if annotation_type is str:
-        return {"type": "string"}
-    if annotation_type is int:
-        return {"type": "integer"}
-    if annotation_type is float:
-        return {"type": "number"}
-    if annotation_type is bool:
-        return {"type": "boolean"}
-    if annotation_type is datetime:
-        return {
-            "type": "string",
-            "format": "date-time",
-        }
-    annotation_type_origin = get_origin(annotation_type)
-    annotation_type_args = get_args(annotation_type)
-    is_union_type = annotation_type_origin is Union
-    if is_union_type:
-        jsonschema_types = []
-        for type_ in annotation_type_args:
-            jsonschema_types.append(_get_jsonschema_type(type_))
-        return {"anyOf": jsonschema_types}
-    is_literal_type = annotation_type_origin is Literal
-    if is_literal_type:
-        enum_values = list(annotation_type_args)
-        unique_enum_types = dict.fromkeys(type(value) for value in enum_values)
-        jsonschema_types = [_get_jsonschema_type(value_type) for value_type in unique_enum_types]
-        result = {}
-        if len(jsonschema_types) == 1:
-            result.update(jsonschema_types[0])
-        elif len(jsonschema_types) > 1:
-            result["anyOf"] = jsonschema_types
-        result["enum"] = enum_values
-        return result
-    is_list_type = (
-        annotation_type_origin is list or annotation_type_origin is collections.abc.Sequence
-    )
-    if is_list_type:
-        result = {"type": "array"}
-        if len(annotation_type_args) == 1:
-            list_item_type = annotation_type_args[0]
-            result["items"] = _get_jsonschema_type(list_item_type)
-        return result
-    is_tuple_type = annotation_type_origin is tuple
-    if is_tuple_type:
-        result = {"type": "array"}
-        if len(annotation_type_args) == 2 and annotation_type_args[-1] is Ellipsis:
-            item_type = annotation_type_args[0]
-            result["items"] = _get_jsonschema_type(item_type)
-        elif annotation_type_args:
-            items = []
-            for arg_type in annotation_type_args:
-                item_schema = _get_jsonschema_type(arg_type)
-                items.append(item_schema)
-            result["items"] = items
-            result["minItems"] = len(annotation_type_args)
-            result["maxItems"] = len(annotation_type_args)
-        return result
-    is_dict_type = (
-        annotation_type_origin is dict or annotation_type_origin is collections.abc.Mapping
-    )
-    if is_dict_type:
-        result = {"type": "object"}
-        if len(annotation_type_args) == 2:
-            # jsonschema requires that the keys in object type are strings, so
-            # we ignore the key type
-            _, value_type = annotation_type_args
-            result["additionalProperties"] = _get_jsonschema_type(value_type)
-        return result
-    is_typed_dict_type = isinstance(annotation_type, _TypedDictMeta)
-    if is_typed_dict_type:
-        result = {"type": "object"}
-        properties = {}
-        for field_name, field_type in annotation_type.__annotations__.items():
-            properties[field_name] = _get_jsonschema_type(field_type)
-        result["properties"] = properties
-        return result
-    if (
-        pydantic is not None
-        and isinstance(annotation_type, type)
-        and issubclass(annotation_type, pydantic.BaseModel)
-    ):
-        return annotation_type.schema()  # type: ignore[no-any-return,attr-defined]
-    return {}
-
-
-# span kinds
-TOOL = OpenInferenceSpanKindValues.TOOL.value
 
 # span attributes
 INPUT_MIME_TYPE = SpanAttributes.INPUT_MIME_TYPE
